Added Services + several changes
